{"types":"C","runs":97}

+++ Assignment2Pyramid.java
@@ -14,8 +14,8 @@ public class Assignment2Pyramid extends GraphicsProgram {
 	private void addBase() {
 	for (int i = 14; i >= 1; i--) {
 		for (int j = 1; j<=BRICKS_IN_BASE; j++) {   	
-				double x = getWidth()/2 - BRICK_WIDTH * BRICKS_IN_BASE/2 + (i * BRICK_WIDTH )   ;
-				double y = getHeight() - j* BRICK_HEIGHT;
+				int x = getWidth()/2 - BRICK_WIDTH * BRICKS_IN_BASE/2 + (i * BRICK_WIDTH )   ;
+				int y = getHeight() - j* BRICK_HEIGHT;
 				GRect brick = new GRect(x, y, BRICK_WIDTH, BRICK_HEIGHT);
 				add(brick);
 			}
